[XPU] Fix int32 overflow in adaptive_avg_pool2d backward for large tensors
AdaptiveAvgPool2dBwdKernelFunctor and AdaptiveAvgPool2dBwdSLMKernelFunctor
stored the input dims (ib_/ic_/ih_/iw_/oh_/ow_) and global_range_ as 32-bit
int, so numel_ = ib_ * ic_ * ih_ * iw_ was computed in 32-bit and overflowed
for tensors with numel > 2^31. When the wrapped product is negative, the
grad-input loop 'for (int64_t i = gi; i < numel_; ...)' runs zero iterations
and grad_input is left entirely unwritten.

Widen the dimension and range members to int64_t so the product is computed in
64-bit. The accessors are already PackedTensorAccessor64, so no addressing
changes are needed. Mirrors the fix pattern used for max_pool3d (#3632).

Verified on Arc Pro B60 (from-source XPU build): with input (1,4,32768,16385)
(numel 2,147,614,720 > 2^31), adaptive_avg_pool2d backward wrote 0/numel
elements before the fix (numel_ wrapped to -2,147,352,576) and all elements
correctly after.

Fixes #4327.

+++ src/ATen/native/xpu/sycl/AdaptiveAveragePooling2dKernels.cpp
@@ -106,14 +106,17 @@ struct AdaptiveAvgPool2dBwdKernelFunctor {
   }
 
  private:
-  int ib_;
-  int ic_;
-  int ih_;
-  int iw_;
-  int oh_;
-  int ow_;
+  // 64-bit dims so `numel_ = ib_ * ic_ * ih_ * iw_` is computed in 64-bit; for
+  // large tensors the int32 product overflowed, truncating numel_ and leaving
+  // the tail of grad_input unwritten.
+  int64_t ib_;
+  int64_t ic_;
+  int64_t ih_;
+  int64_t iw_;
+  int64_t oh_;
+  int64_t ow_;
   int64_t numel_;
-  int global_range_;
+  int64_t global_range_;
   int local_range_;
   PackedTensorAccessor64<const scalar_t, 4> gyacc_;
   PackedTensorAccessor64<scalar_t, 4> gxacc_;
@@ -223,15 +226,18 @@ struct AdaptiveAvgPool2dBwdSLMKernelFunctor
   }
 
  private:
-  int ib_;
-  int ic_;
-  int ih_;
-  int iw_;
-  int oh_;
-  int ow_;
+  // 64-bit dims so `numel_ = ib_ * ic_ * ih_ * iw_` is computed in 64-bit; for
+  // large tensors the int32 product overflowed, truncating numel_ and leaving
+  // the tail of grad_input unwritten.
+  int64_t ib_;
+  int64_t ic_;
+  int64_t ih_;
+  int64_t iw_;
+  int64_t oh_;
+  int64_t ow_;
   int64_t numel_;
   int local_range_;
-  int global_range_;
+  int64_t global_range_;
   PackedTensorAccessor64<const scalar_t, 4> gyacc_;
   PackedTensorAccessor64<scalar_t, 4> gxacc_;
   sycl_local_acc_t<int> _oh0_cached_;
